// projects/app/src/app/models/info.ts
export interface info{
    server_version: string
    chain_id: string
    head_block_num: number
    head_block_id: string
    head_block_time: string
    head_block_producer: string
    last_irreversible_block_num: number
    last_irreversible_block_id: string
    virtual_block_cpu_limit: number
    virtual_block_net_limit: number
    block_cpu_limit: number
    block_net_limit: number
    server_version_string: string
    fork_db_head_block_num: number
    fork_db_head_block_id: string
    server_full_version_string: string
  }